Revert "Make credit2 the default scheduler"
The migration code reads scheduler parameters on the sending side and
tries to set them again on the receiving side, failing if this fails;
the result is that a simple upgrade from 4.11 -> 4.12 will now fail
all migrations.  Solving this is not simple; revert the credit2
upgrade until we can sort something out.

This reverts commit ec6662445b1ec9231ca69583863de644206ef313.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>