tools: Once again honour, but no longer advertise GIT_HTTP env var
In "build: add autoconf to replace custom checks in tools/check"
--enable-githttp was introduced.  But we missed this comment where it
was advertised.

Also, that commit had the effect of uncondtionally setting GIT_HTTP
from the configure variable.  But the env var has been advertised in
some places as the way to specify this behaviour, and overriding it is
just unfriendly.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>
CC: Paul Durrant <paul.durrant@citrix.com>
Acked-by: Wei Liu <wei.liu2@citrix.com>